+++ index.js
@@ -68,7 +68,7 @@ const itemDetailController = require('./controllers/web/itemDetail')
 const storeItemsController = require('./controllers/web/storeItems')
 const itemsListController = require('./controllers/web/itemList')
 const itemsSaveController = require('./controllers/web/saveItems')
-const getUserStatsController = require('./controllers/web/getUserStats')
+const getUserStatsController = require('./controllers/web/get-user-stats')
 const getUserStatsSubmitController = require('./controllers/web/getUserStatsSubmit')
 const getChartsController = require('./controllers/web/get-charts-controller')
 const getChartsOverallController = require('./controllers/web/get-charts-overall-controller')
